add upload album title into database
